11章の実装

+++ src/app/admin/categories/[id]/page.tsx
@@ -2,35 +2,43 @@
 // 管理者_カテゴリーの編集ページ
 "use client";
 
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useParams, useRouter } from "next/navigation";
 import { CategoryForm } from "../_components/CategoryForm";
 import { UpdateCategoryRequestBody } from "@/app/api/admin/categories/[id]/route";
 import { CategoryShowResponse } from "./_types/CategoryShowResponse";
+import { useSupabaseSession } from "../../../_hooks/useSupabaseSession";
+import { useForm, FieldErrors } from "react-hook-form";
 
 export default function AdminEditCategory() {
-  const [name, setName] = useState("");
-  const [isSubmitting, setIsSubmitting] = useState(false);
   const { id } = useParams();
   const router = useRouter();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    // フォームのデフォルトの動作をキャンセルします。
-    e.preventDefault();
+  const { token } = useSupabaseSession();
+  // 既定値を準備
+  const defaultValues = {
+    name: "",
+  };
+  // フォームを初期化
+  const {
+    register,
+    handleSubmit,
+    formState: { errors, isSubmitting },
+    reset,
+  } = useForm({
+    defaultValues,
+  });
+  // サブミット時の処理
+  const onSubmit = async (data: UpdateCategoryRequestBody) => {
+    if (!token) return;
 
     try {
-      setIsSubmitting(true);
-
-      const body: UpdateCategoryRequestBody = {
-        name,
-      };
-
       await fetch(`/api/admin/categories/${id}`, {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
+          Authorization: token,
         },
-        body: JSON.stringify(body),
+        body: JSON.stringify(data),
       });
 
       alert("カテゴリーを更新しました。");
@@ -39,18 +47,23 @@ export default function AdminEditCategory() {
     } catch (error) {
       console.error("カテゴリーの更新に失敗しました:", error);
       alert("カテゴリーの更新に失敗しました。");
-    } finally {
-      setIsSubmitting(false);
     }
   };
 
-  const handleDeletePost = async () => {
+  const onError = (err: FieldErrors<FormData>) => console.log(err);
+
+  const onDelete = async () => {
     if (!confirm("カテゴリーを削除しますか？")) return;
 
+    if (!token) return;
+
     try {
-      setIsSubmitting(true);
       await fetch(`/api/admin/categories/${id}`, {
         method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
       });
 
       alert("カテゴリーを削除しました");
@@ -59,20 +72,26 @@ export default function AdminEditCategory() {
     } catch (error) {
       console.error("カテゴリーの削除に失敗しました。", error);
       alert("カテゴリーの削除に失敗しました。");
-    } finally {
-      setIsSubmitting(false);
     }
   };
 
   useEffect(() => {
+    if (!token) return;
+
     const fetcher = async () => {
-      const res = await fetch(`/api/admin/categories/${id}`);
+      const res = await fetch(`/api/admin/categories/${id}`, {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+      });
       const { category }: { category: CategoryShowResponse["category"] } =
         await res.json();
-      setName(category.name);
+
+      reset({ name: category.name });
     };
     fetcher();
-  }, [id]);
+  }, [id, token, reset]);
 
   return (
     <div className="">
@@ -81,11 +100,11 @@ export default function AdminEditCategory() {
       </div>
       <CategoryForm
         mode="edit"
-        name={name}
-        setName={setName}
-        onSubmit={handleSubmit}
-        onDelete={handleDeletePost}
+        onSubmit={handleSubmit(onSubmit, onError)}
+        onDelete={onDelete}
         disabled={isSubmitting}
+        register={register}
+        errors={errors}
       />
     </div>
   );
