+++ src/app/admin/posts/[id]/page.tsx
@@ -2,44 +2,49 @@
 // 管理者_記事の編集ページ
 "use client";
 
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useParams, useRouter } from "next/navigation";
 import { PostForm } from "../_components/PostForm";
-import {
-  Category,
-  UpdatePostRequestBody,
-} from "@/app/api/admin/posts/[id]/route";
+import { UpdatePostRequestBody } from "@/app/api/admin/posts/[id]/route";
 import { PostShowResponse } from "./_types/PostShowResponse";
+import { useSupabaseSession } from "../../../_hooks/useSupabaseSession";
+import { useForm, FieldErrors } from "react-hook-form";
 
 export default function AdminEditPost() {
-  const [title, setTitle] = useState("");
-  const [content, setContent] = useState("");
-  const [thumbnailUrl, setThumbnailUrl] = useState("");
-  const [categories, setCategories] = useState<Category[]>([]);
-  const [isSubmitting, setIsSubmitting] = useState(false);
   const { id } = useParams();
   const router = useRouter();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    // フォームのデフォルトの動作をキャンセルします。
-    e.preventDefault();
+  const { token } = useSupabaseSession();
+  // 既定値を準備
+  const defaultValues: UpdatePostRequestBody = {
+    title: "",
+    content: "",
+    thumbnailImageKey: "",
+    categories: [],
+  };
+  // フォームを初期化
+  const {
+    register,
+    handleSubmit,
+    control,
+    formState: { errors, isSubmitting },
+    reset,
+    setValue,
+    watch,
+  } = useForm<UpdatePostRequestBody>({
+    defaultValues,
+  });
+  // サブミット時の処理
+  const onSubmit = async (data: UpdatePostRequestBody) => {
+    if (!token) return;
 
     try {
-      setIsSubmitting(true);
-
-      const body: UpdatePostRequestBody = {
-        title,
-        content,
-        thumbnailUrl,
-        categories,
-      };
-
       await fetch(`/api/admin/posts/${id}`, {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
+          Authorization: token,
         },
-        body: JSON.stringify(body),
+        body: JSON.stringify(data),
       });
 
       alert("記事を更新しました。");
@@ -48,18 +53,23 @@ export default function AdminEditPost() {
     } catch (error) {
       console.error("記事の更新に失敗しました:", error);
       alert("記事の更新に失敗しました。");
-    } finally {
-      setIsSubmitting(false);
     }
   };
 
-  const handleDeletePost = async () => {
+  const onError = (err: FieldErrors<UpdatePostRequestBody>) => console.log(err);
+
+  const onDelete = async () => {
     if (!confirm("記事を削除しますか？")) return;
 
+    if (!token) return;
+
     try {
-      setIsSubmitting(true);
       await fetch(`/api/admin/posts/${id}`, {
         method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
       });
 
       alert("記事を削除しました");
@@ -68,22 +78,32 @@ export default function AdminEditPost() {
     } catch (error) {
       console.error("記事の削除に失敗しました。", error);
       alert("記事の削除に失敗しました。");
-    } finally {
-      setIsSubmitting(false);
     }
   };
 
   useEffect(() => {
+    if (!token) return;
+
     const fetcher = async () => {
-      const res = await fetch(`/api/admin/posts/${id}`);
+      const res = await fetch(`/api/admin/posts/${id}`, {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+      });
       const { post }: { post: PostShowResponse["post"] } = await res.json();
-      setTitle(post.title);
-      setContent(post.content);
-      setThumbnailUrl(post.thumbnailUrl);
-      setCategories(post.postCategories.map((pc) => pc.category));
+      reset({
+        title: post.title,
+        content: post.content,
+        thumbnailImageKey: post.thumbnailImageKey,
+        categories: post.postCategories.map((pc) => ({
+          id: pc.category.id,
+          name: pc.category.name,
+        })),
+      });
     };
     fetcher();
-  }, [id]);
+  }, [id, token, reset]);
 
   return (
     <div className="">
@@ -92,17 +112,14 @@ export default function AdminEditPost() {
       </div>
       <PostForm
         mode="edit"
-        title={title}
-        setTitle={setTitle}
-        content={content}
-        setContent={setContent}
-        thumbnailUrl={thumbnailUrl}
-        setThumbnailUrl={setThumbnailUrl}
-        categories={categories}
-        setCategories={setCategories}
-        onSubmit={handleSubmit}
-        onDelete={handleDeletePost}
+        onSubmit={handleSubmit(onSubmit, onError)}
+        onDelete={onDelete}
         disabled={isSubmitting}
+        register={register}
+        errors={errors}
+        setValue={setValue}
+        watch={watch}
+        control={control}
       />
     </div>
   );
